+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Verify activity sender instead of actor when filtering spam.
+- Ignore forwarded `Announce` activities.
 
 ## [4.6.0] - 2025-07-16